Extract ObjectDetailHeader to a separate file
