+++ frontend/src/metabase/visualizations/components/ObjectDetail/ObjectDetail.tsx
@@ -7,7 +7,6 @@ import { useMount, usePrevious } from "react-use";
 import { State } from "metabase-types/store";
 import type { ConcreteTableId, DatasetData } from "metabase-types/api";
 
-import Button from "metabase/core/components/Button";
 import { NotFound } from "metabase/containers/ErrorPages";
 import LoadingSpinner from "metabase/components/LoadingSpinner";
 
@@ -47,14 +46,12 @@ import {
 import {
   RootModal,
   ObjectDetailContainer,
-  ObjectDetailHeaderWrapper,
-  ObjectIdLabel,
-  CloseButton,
   ErrorWrapper,
   PaginationFooter,
   ObjectDetailWrapperDiv,
 } from "./ObjectDetail.styled";
 import { ObjectDetailBody } from "./ObjectDetailBody";
+import { ObjectDetailHeader } from "./ObjectDetailHeader";
 
 const mapStateToProps = (state: State, { data }: ObjectDetailProps) => {
   const isLoggedIn = !!getUser(state);
@@ -401,77 +398,6 @@ export function ObjectDetailWrapper({
         />
       )}
     </>
-  );
-}
-
-export interface ObjectDetailHeaderProps {
-  canZoom: boolean;
-  objectName: string;
-  objectId: ObjectId | null | unknown;
-  canZoomPreviousRow: boolean;
-  canZoomNextRow?: boolean;
-  showActions?: boolean;
-  viewPreviousObjectDetail: () => void;
-  viewNextObjectDetail: () => void;
-  closeObjectDetail: () => void;
-}
-
-export function ObjectDetailHeader({
-  canZoom,
-  objectName,
-  objectId,
-  canZoomPreviousRow,
-  canZoomNextRow,
-  showActions = true,
-  viewPreviousObjectDetail,
-  viewNextObjectDetail,
-  closeObjectDetail,
-}: ObjectDetailHeaderProps): JSX.Element {
-  return (
-    <ObjectDetailHeaderWrapper className="Grid">
-      <div className="Grid-cell">
-        <h2 className="p3">
-          {objectName}
-          {objectId !== null && <ObjectIdLabel> {objectId}</ObjectIdLabel>}
-        </h2>
-      </div>
-      {showActions && (
-        <div className="flex align-center">
-          <div className="flex p2">
-            {!!canZoom && (
-              <>
-                <Button
-                  data-testid="view-previous-object-detail"
-                  onlyIcon
-                  borderless
-                  className="mr1"
-                  disabled={!canZoomPreviousRow}
-                  onClick={viewPreviousObjectDetail}
-                  icon="chevronup"
-                />
-                <Button
-                  data-testid="view-next-object-detail"
-                  onlyIcon
-                  borderless
-                  disabled={!canZoomNextRow}
-                  onClick={viewNextObjectDetail}
-                  icon="chevrondown"
-                />
-              </>
-            )}
-            <CloseButton>
-              <Button
-                data-testid="object-detail-close-button"
-                onlyIcon
-                borderless
-                onClick={closeObjectDetail}
-                icon="close"
-              />
-            </CloseButton>
-          </div>
-        </div>
-      )}
-    </ObjectDetailHeaderWrapper>
   );
 }
